+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add checks for Cloud Build email roles in `cr_setup()`
 * Add direct use of Secret Manager environment args in builds via `availableSecrets` (#106)
+* Add support for `logsBucket` in `cr_build_yaml` and `cr_build_make`
 
 # googleCloudRunner 0.4.1